Enforce SDLC control atom separation
